Cleanup and remove state approach
