+++ lib/server.go
@@ -3461,10 +3461,6 @@ func (srv *Server) handleFullySyncedStateAMQP() {
 				srv.eventManager.OnBlockAccepted(func(event *BlockEvent) {
 					go PublishBlockEvent(event, amqpDest)
 				})
-				// srv.eventManager.OnStateSyncerOperation(func(event *StateSyncerOperationEvent) {
-				// 	//Run async
-				// 	go PublishStateChangeEvent(event.StateChangeEntry, amqpDest)
-				// })
 				glog.Infof("AMQP publisher event handling set.")
 			}
 		}
